doc insomnia e acessAuth

+++ back-end/src/controllers/adm.controller.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import Adm from "../models/adm.model.js";
 import { validationError } from "../validatorError/validationError.js";
 
@@ -15,7 +16,16 @@ export const getByIdAdm = async (req, res) => {
   try {
     const { id } = req.params;
 
+      // Verificar se o ID é um ObjectId válido
+    if(!mongoose.Types.ObjectId.isValid(id)){
+      return res.status(400).json({ message: "ID inválido" });
+    }
+
     const adm = await Adm.findById(id);
+
+    if (!adm) {
+      return res.status(404).json({ message: "Administrador não encontrado" });
+    }
 
     return res.status(200).json(adm);
   } catch (error) {
@@ -26,8 +36,8 @@ export const getByIdAdm = async (req, res) => {
 export const createAdm = async (req, res) => {
   try {
     const adm = req.body;
-
-    const admCreate = await Adm.create(data);
+  
+    const admCreate = await Adm.create(adm);
 
     return res.status(200).json(admCreate);
   } catch (error) {
@@ -38,10 +48,18 @@ export const createAdm = async (req, res) => {
 export const updateAdm = async (req, res) => {
   try {
     const { id } = req.params;
+   
+    if(!mongoose.Types.ObjectId.isValid(id)){
+      return res.status(400).json({ message: "ID inválido" });
+    }
 
     const data = req.body;
 
     const adm = await Adm.findByIdAndUpdate(id, data, { new: true });
+
+   if (!adm) {
+      return res.status(404).json({ message: "Administrador não encontrado" });
+    }
 
     return res.status(200).json(adm);
   } catch (error) {
@@ -52,7 +70,16 @@ export const updateAdm = async (req, res) => {
 export const removeAdm = async (req, res) => {
   try {
     const { id } = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+      return res.status(400).json({ message: "ID inválido" });
+    }
+
     const deleteAdm = await Adm.findByIdAndDelete(id);
+
+    if (!deleteAdm) {
+      return res.status(404).json({ message: "Administrador não encontrado" });
+    }
 
     return res.status(200).json(deleteAdm);
   } catch (error) {
