+++ back-end/src/controllers/collaborator.controller.js
@@ -1,3 +1,4 @@
+import mongoose from "mongoose";
 import Collaborator from "../models/collaborator.model.js";
 import { validationError } from "../validatorError/validationError.js";
 
@@ -14,7 +15,15 @@ export const getCollaboratorId = async (req, res) => {
   try {
     const { id } = req.params;
 
+    if(!mongoose.Types.ObjectId.isValid(id)){
+      return res.status(400).json({ message: "ID inválido" });
+    }
+
     const collaborator = await Collaborator.findById(id);
+
+    if(!collaborator){
+      return res.status(404).json({ message: "Colaborador não encontrado" });
+    }
 
     return res.status(200).json(collaborator);
   } catch (error) {
@@ -37,11 +46,20 @@ export const createCollaborator = async (req, res) => {
 export const updateCollabarator = async (req, res) => {
   try {
     const { id } = req.params;
+
+    if(!mongoose.Types.ObjectId.isValid(id)){
+      return res.status(400).json({ message: "ID inválido" });
+    }
+
     const data = req.body;
 
     const collaborator = await Collaborator.findByIdAndUpdate(id, data, {
       new: true,
     });
+
+    if(!collaborator){
+      return res.status(404).json({ message: "Colaborador não encontrado" });
+    }
 
     return res.status(200).json(collaborator);
   } catch (error) {
@@ -52,8 +70,17 @@ export const updateCollabarator = async (req, res) => {
 export const removeCollaborator = async (req, res) => {
   try {
     const { id } = req.params;
+   
+    if(!mongoose.Types.ObjectId.isValid(id)){
+      return res.status(400).json({ message: "ID inválido" });
+    }
 
     const deleteCollaborator = await Collaborator.findByIdAndDelete(id);
+
+    
+    if(!deleteCollaborator){
+      return res.status(404).json({ message: "Colaborador não encontrado" });
+    }
 
     return res.status(200).json(deleteCollaborator);
   } catch (error) {
